Enable boot/bootloader/radio version checks
This enables the Build.isBuildConsistent checks after tweaking the radio
check to stop failing when the radio hasn't been turned on. The property
used to check the radio version (gsm.baseband.version) is only set once
the radio is turned on and initialized.

The API documentation already claims that this is done.

Change-Id: If4e60ed1254b73cc85c22ff8ec86b15446de771c
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>